refactor: Clean up utility functions
