+++ config/presets/storage/backend/storage-memory.json
@@ -2,17 +2,10 @@
   "@context": "https://linkedsoftwaredependencies.org/bundles/npm/@solid/community-server/^1.0.0/components/context.jsonld",
   "@graph": [
     {
-      "@id": "urn:solid-server:default:MetadataController",
-      "@type": "MetadataController"
-    },
-    {
       "@id": "urn:solid-server:default:MemoryDataAccessor",
       "@type": "InMemoryDataAccessor",
       "InMemoryDataAccessor:_base": {
         "@id": "urn:solid-server:default:variable:base"
-      },
-      "InMemoryDataAccessor:_metadataController": {
-        "@id": "urn:solid-server:default:MetadataController"
       }
     },
     {
@@ -23,12 +16,6 @@
       },
       "DataAccessorBasedStore:_base": {
         "@id": "urn:solid-server:default:variable:base"
-      },
-      "DataAccessorBasedStore:_metadataController": {
-        "@id": "urn:solid-server:default:MetadataController"
-      },
-      "DataAccessorBasedStore:_containerManager": {
-        "@id": "urn:solid-server:default:UrlContainerManager"
       }
     }
   ]
