+++ src/ldp/http/BasicResponseWriter.ts
@@ -2,7 +2,7 @@ import { getLoggerFor } from '../../logging/LogUtil';
 import type { HttpResponse } from '../../server/HttpResponse';
 import { INTERNAL_QUADS } from '../../util/ContentTypes';
 import { UnsupportedHttpError } from '../../util/errors/UnsupportedHttpError';
-import { pipeSafe } from '../../util/Util';
+import { pipeSafely } from '../../util/StreamUtil';
 import type { MetadataWriter } from './metadata/MetadataWriter';
 import type { ResponseDescription } from './response/ResponseDescription';
 import { ResponseWriter } from './ResponseWriter';
@@ -34,7 +34,7 @@ export class BasicResponseWriter extends ResponseWriter {
     input.response.writeHead(input.result.statusCode);
 
     if (input.result.data) {
-      const pipe = pipeSafe(input.result.data, input.response);
+      const pipe = pipeSafely(input.result.data, input.response);
       pipe.on('error', (error): void => {
         this.logger.error(`Writing to HttpResponse failed with message ${error.message}`);
       });
